requirements.txt
awsebcli==3.14.1
blessed==1.15.0
boto3==1.7.45
botocore==1.10.45
cached-property==1.4.3
cairocffi==0.8.0
CairoSVG==2.1.3
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.4
django-braces==1.12.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-model-utils==3.1.1
django-registration-redux==2.1
django-storages==1.6.6
django-webpack-loader==0.5.0
djangorestframework==3.7.7
docker==3.4.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
html5lib==1.0.1
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
Markdown==2.6.11
mysqlclient==1.3.12
pathspec==0.5.5
pdfrw==0.4
Pillow==5.1.0
pycparser==2.18
Pyphen==0.9.4
python-dateutil==2.7.3
pytz==2018.7
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
sqlparse==0.2.4
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
tinycss2==0.6.1
urllib3==1.22
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
websocket-client==0.48.0
